release 0.2.3 with public Rust XLIL APIs and contextual Bool lowering

+++ xslang/sources/mir/model.rs
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+ * SPDX-FileCopyrightText: 2026 Leitwolf <support@xsharp-lang.xyz>
  * SPDX-License-Identifier: MPL-2.0
  */
 
@@ -95,7 +95,7 @@ pub enum Statement
   ConstStr
   {
     local: LocalId,
-    units: Vec<u16>,
+    units: Vec<u32>,
     span: Span,
   },
   ConstBool
